Minor changes in code structure

+++ app/implementation.py
@@ -1,32 +1,47 @@
 from dbquery import querydb
+from schema import Schema, And, Use
 
 
 def getAllFlowers():
-
     data = 'SELECT * FROM flower'
     return querydb(data, operation='GET', check='list')
 
 
 def postFlower(request):
+    data = "INSERT INTO flower (name_ser, name_lat, description, watering_period) values ('" + str(
+        request.json['name_ser']) + "', '" + str(request.json['name_lat']) + "', '" + str(
+        request.json['description']) + "', '" + str(request.json['watering_period']) + "')"
 
-    return querydb("", operation='POST', request=request)
+    return querydb(data, operation='POST')
 
 
 def getFlowerByID(flower_id):
-
     data = 'SELECT * FROM flower where flower_id = ' + str(flower_id)
     return querydb(data, 'GET', 'tuple', flower_id=flower_id)
 
 
 def putFlowerByID(request, flower_id):
-
-    return querydb("", 'PUT', flower_id=flower_id, request=request)
+    data = getFlowerByID(flower_id)
+    if data == "No data to return.":
+        return postFlower(request)
+    else:
+        putData = putDataCheck(request, data)
+        if putData == "Something went wrong in mapping data.":
+            return {"msg": "Something went wrong in mapping data."}, 500
+        data = "UPDATE flower SET flower_id = '" + str(flower_id) + "', name_ser = '" + putData[0] + "', name_lat = '" + \
+               putData[1] + "', description = '" + putData[2] + "', watering_period = '" + putData[
+                   3] + "' WHERE flower_id = '" + str(flower_id) + "'"
+        return querydb(data, 'PUT', flower_id=flower_id)
 
 
 def deleteFlowerByID(flower_id):
 
-    data = 'DELETE FROM flower WHERE flower_id = ' + (str(flower_id))
-    return querydb(data, 'DELETE', flower_id=flower_id)
+    data = getFlowerByID(flower_id)
+    if data == "No data to return.":
+        return {"msg": "Flower with flower_id " + str(flower_id) + " does not exist in DB."}
+    else:
+        data = 'DELETE FROM flower WHERE flower_id = ' + (str(flower_id))
+        return querydb(data, 'DELETE', flower_id=flower_id)
 
 
 def putDataCheck(request, data):
@@ -52,3 +67,9 @@ def putDataCheck(request, data):
     except:
         return "Something went wrong in mapping data."
 
+
+def getSchema():
+    return Schema({'name_ser': And(str, len),
+                   'name_lat': And(str, len),
+                   'description': And(str, len),
+                   'watering_period': And(Use(int))})
